+++ extensions/applicationinsights-properties-js/src/Context/Internal.ts
@@ -4,7 +4,7 @@
 import { IInternal } from '@microsoft/applicationinsights-common';
 import { ITelemetryConfig } from '../Interfaces/ITelemetryConfig';
 
-const Version = "2.5.10";
+const Version = "2.5.11";
 
 export class Internal implements IInternal {