+++ shared/AppInsightsCore/rollup.config.js
@@ -1,6 +1,6 @@
-import nodeResolve from "rollup-plugin-node-resolve";
-import {uglify} from "rollup-plugin-uglify";
-import replace from "rollup-plugin-replace";
+import nodeResolve from "@rollup/plugin-node-resolve";
+import { uglify } from "@microsoft/applicationinsights-rollup-plugin-uglify3-js";
+import replace from "@rollup/plugin-replace";
 import dynamicRemove from "@microsoft/dynamicproto-js/tools/rollup/node/removedynamic";
 import { es3Poly, es3Check, importCheck } from "@microsoft/applicationinsights-rollup-es3";