+++ tools/rollup-plugin-uglify3-js/src/uglify3-js.ts
@@ -1,0 +1,58 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import * as UglifyJs from 'uglify-js';
+
+function isSourceMapEnabled(options) {
+    if (options) {
+        return options.sourceMap !== false && options.sourcemap !== false;
+    }
+
+    return false;
+} 
+
+function _doMinify(code, filename, options, chunkOptions) {
+    var theCode = {};
+    theCode[filename] = code;
+
+    let assign = Object['assign'];
+    let theOptions = assign({}, options);
+    if (theOptions.hasOwnProperty("sourcemap")) {
+        delete theOptions.sourcemap;
+    }
+
+    if (isSourceMapEnabled(options)) {
+        theOptions.sourceMap = {
+            filename: filename
+        }
+        if (filename) {
+            theOptions.sourceMap.url = filename + ".map";
+        }
+    }
+
+    var result = UglifyJs.minify(theCode, theOptions);
+
+    if (result.error) {
+        throw new Error(JSON.stringify(result.error));
+    }
+
+    var transform:any = {
+        code: result.code
+    };
+
+    if (isSourceMapEnabled(options) && result.map) {
+        transform.map = result.map;
+    }
+
+    return transform;
+}
+
+export function uglify(options = {}) {
+
+    return {
+        name: 'ai-rollup-plugin-uglify3-js',
+        renderChunk(code, chunk, chkOpt) {
+            return _doMinify(code, chunk.filename, options, chkOpt);
+        }
+    }
+}