+++ RELEASES.md
@@ -1,5 +1,18 @@
 # Releases
 
+## 2.5.11 (January 15th, 2021)
+
+### Changelog
+
+- #1452 [BUG] v2.5.10 Snippet Initialization fails to redirect proxied functions -- causing terminal exception
+
+### Update Click Analytics plugin to v2.5.11
+
+- Update Core dependency to ^2.5.11 Core changes
+- #1441 [BUG] Fix issues based on pageName,clickevent name and improved the way we collect useful telemetry data
+- Updated Docs
+
+
 ## 2.5.10 (November 16th, 2020)
 
 ### New extension @microsoft/applicationinsights-clickanalytics-js